+++ react-materials/scaffolds/ice-design-pro/src/pages/GeneralTable/components/ContractTable/ContractTable.jsx
@@ -35,11 +35,11 @@ export default class ContractTable extends Component {
     return (
       <div>
         <a style={styles.link} onClick={this.handleUpdate}>
-          修改合同
+          修改
         </a>
         <span style={styles.separator} />
         <a style={styles.link} onClick={this.handleMore}>
-          查看详情
+          查看
         </a>
       </div>
     );
@@ -63,12 +63,6 @@ export default class ContractTable extends Component {
         title: '我方公司',
         dataIndex: 'ourCompany',
         key: 'ourCompany',
-        width: 160,
-      },
-      {
-        title: '对方公司',
-        dataIndex: 'otherCompany',
-        key: 'otherCompany',
         width: 160,
       },
       {