+++ tools/ice-devtools/config/getBabelConfig.js
@@ -48,6 +48,10 @@ module.exports = (buildConfig = {}) => {
       ],
       ['babel-plugin-import', { libraryName: '@alife/next' }, '@alife/next'],
       ['babel-plugin-import', { libraryName: '@alifd/next' }, '@alifd/next'],
+      ['babel-plugin-import', { libraryName: 'antd', style: true }, 'antd'],
+      [
+        '@babel/plugin-transform-runtime'
+      ]
     ]),
   };
 };