+++ tools/ice-devtools/lib/component/build.js
@@ -2,7 +2,7 @@ const { createReadStream, createWriteStream, writeFileSync } = require('fs');
 const babel = require('@babel/core');
 const chokidar = require('chokidar');
 const colors = require('colors');
-const dtsGenerator = require('typescript-definition-generator');
+const dtsGenerator = require('@alifd/dts-generator');
 const glob = require('glob');
 const mkdirp = require('mkdirp');
 const path = require('path');
@@ -10,7 +10,10 @@ const propsSchemaGenerator = require('props-schema-generator');
 const rimraf = require('rimraf');
 
 const getBabelConfig = require('../../config/getBabelConfig');
+const getBaseConfig = require('../../config/webpack.component');
 const ComponentStyleGenerator = require('../../utils/component-style-generator');
+const buildCombinedDemo = require('../../utils/build-combined-demo');
+const info = require('./info');
 
 const GLOB_PATTERN = '**/*';
 const babelOpt = getBabelConfig();
@@ -19,6 +22,23 @@ const babelOpt = getBabelConfig();
  * 构建项目
  */
 module.exports = function componentBuild(workDir, opts) {
+  const config = getBaseConfig(workDir);
+
+  if (process.env.SKIP_DEMO) {
+    compile(workDir, opts);
+    return;
+  }
+
+  // HACK：放在回调中执行，是为了避免两个任务的 log 信息混在一起
+  buildCombinedDemo(workDir, config, (err) => {
+    if (!err) {
+      info(workDir);
+      compile(workDir, opts);
+    }
+  });
+}
+
+function compile(workDir, opts) {
   opts = opts || {};
 
   const srcDir = path.join(workDir, 'src');
@@ -88,6 +108,8 @@ module.exports = function componentBuild(workDir, opts) {
         writeFileSync(dtsDist, dts.message);
         console.log(colors.green('Write index.d.ts'));
       }
+    }).catch((err) => {
+      console.log(colors.yellow('生成 d.ts 失败'), err);
     });
   }