+++ tools/ice-devtools/CHANGELOG.md
@@ -1,15 +1,71 @@
-# ice-scripts Changelog
+# Changelog
+
+## 2.1.20
+ - [feat] 引入 babel-plugin-transform-runtime
+ - [feat] 支持编译 ts
+ - [fix] markdown 中 code 解析时的默认 lang
+ - [fix] README 中代码片段被 slice 的问题
+
+## 2.1.19
+ - [feat] 支持 antd 组件样式加载
+ - [feat] dev 支持 ip 访问
+
+## 2.1.18
+ - [fix] homepage url 随 npm 版本更新
+
+## 2.1.17
+ - [feat] 增加组件 homepage 字段
+
+## 2.1.16
+ - [feat] 增加 markdown 中代码片段支持类型
+
+## 2.1.15
+ - [feat] 增加组件 demo 构建逻辑
+
+## 2.1.14
+ - [feat] 增加埋点
+ - [feat] 更新提醒
+ - [feat] 添加组件入库托管
+
+## 2.1.13
+
+- [fix] 区块 external React 后预览错误
+
+## 2.1.12
+
+- [fix] 修改 npm registry 为官方默认值
+
+## 2.1.10
+
+- [chore] 组件和区块开发不 external React
+
+## 2.1.9
+
+- [feat] 添加内部站点物料同步
+
+## 2.1.8
+
+- [fix] 修复 imagemin-pngquant 安装报错 [#46](https://github.com/imagemin/imagemin-pngquant/issues/46)
+
+
+## 2.1.7
+
+- [chore] typescript-definition-generator 迁移为 @alifd/dts-generator [#1475](https://github.com/alibaba/ice/pull/1475)
 
 ## 2.1.6
+
 - [fix] 修复 block build 报错 [#1444](https://github.com/alibaba/ice/issues/1444)
 
 ## 2.1.5
+
 - [fix] 增加默认 screenshot
 
 ## 2.1.4
+
 - [feat] 兼容 ice 仓库的目录结构
 
 ## 2.1.3
+
 - [feat] 支持 CSS Modules
 
 ## 2.1.2